[OPIK-6303] [FE] refactor: extract feedback score conditions into a shared component
